# openapi.json
{"openapi":"3.0.2","info":{"title":"Orbis 2 Webservice","version":"1.0"},"paths":{"/addDocumentToAnnotationQueue/{da_id}":{"put":{"summary":"Add Document To Annotation Queue","operationId":"add_document_to_annotation_queue_addDocumentToAnnotationQueue__da_id__put","parameters":[{"required":true,"schema":{"title":"Da Id","type":"integer"},"name":"da_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/getDocumentForAnnotation":{"get":{"summary":"Get Document For Annotation","operationId":"get_document_for_annotation_getDocumentForAnnotation_get","parameters":[{"required":false,"schema":{"title":"Corpus Name"},"name":"corpus_name","in":"query"},{"required":false,"schema":{"title":"Annotator"},"name":"annotator","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/getDocumentContent":{"get":{"summary":"Get Document Content","operationId":"get_document_content_getDocumentContent_get","parameters":[{"required":false,"schema":{"title":"Da Id"},"name":"da_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/getDocumentAnnotations":{"get":{"summary":"Get Document Annotations","operationId":"get_document_annotations_getDocumentAnnotations_get","parameters":[{"required":false,"schema":{"title":"Da Id"},"name":"da_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/saveDocumentAnnotations":{"post":{"summary":"Save Document Annotations","operationId":"save_document_annotations_saveDocumentAnnotations_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataExchangeModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/addDocument":{"post":{"summary":"Add Doccument","operationId":"add_doccument_addDocument_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentPostModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/createCorpus":{"post":{"summary":"Create Corpus","operationId":"create_corpus_createCorpus_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CorpusModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnnotationBlobModel":{"title":"AnnotationBlobModel","required":["d_id","meta","annotations"],"type":"object","properties":{"d_id":{"title":"D Id","type":"string","description":"Unique key for identifying document"},"meta":{"title":"Meta","type":"object"},"annotations":{"title":"Annotations","type":"array","items":{"$ref":"#/components/schemas/AnnotationModel"}}},"example":{"d_id":"some_document_id","meta":{"some":"annotator","and":"iteration","or":"creation_date"},"annotations":[{"key":"#somekey","type":"topic","surface_form":"Hello world","start":21,"end":32,"scope":"some_scope","meta":{"some":"other","key":"values"}}]}},"AnnotationModel":{"title":"AnnotationModel","required":["key","type","surface_form","start","end","scope","meta"],"type":"object","properties":{"key":{"title":"Key","type":"string","description":"Key which uniquely assigns the element to the corresponding entry in the x28 ontology."},"type":{"title":"Type","type":"string","description":"Annotation type."},"surface_form":{"title":"Surface Form","type":"string","description":"Annotated element as it is written in the text representation of the html."},"start":{"title":"Start","type":"integer","description":"Start position of the annotated element in the text representation of the html."},"end":{"title":"End","type":"integer","description":"Start position of the annotated element in the text representation of the html."},"scope":{"title":"Scope","type":"string","description":"Scope of the annotations"},"meta":{"title":"Meta","type":"object"}},"example":{"key":"#somekey","type":"topic","surface_form":"Hello world","start":21,"end":32,"scope":"some_scope","meta":{"some":"other","key":"values"}}},"CorpusModel":{"title":"CorpusModel","required":["corpus_name","description"],"type":"object","properties":{"corpus_name":{"title":"Corpus Name","type":"string","description":"Name of the corpus."},"description":{"title":"Description","type":"string","description":"Description of the corpus."}},"example":{"corpus_name":"Some corpus","description":"Just a random corpus"}},"DataExchangeModel":{"title":"DataExchangeModel","required":["da_id","annotator","data"],"type":"object","properties":{"da_id":{"title":"Da Id","type":"string","description":"Unique key for identifying document annotations"},"annotator":{"title":"Annotator","type":"string","description":"Name of the annotator"},"data":{"$ref":"#/components/schemas/AnnotationBlobModel"}},"example":{"da_id":"some_da_id","annotator":"some_annotator","data":{"d_id":"some_document_id","meta":{"some":"annotator","and":"iteration","or":"creation_date"},"annotations":[{"key":"#somekey","type":"topic","surface_form":"Hello world","start":21,"end":32,"scope":"some_scope","meta":{"some":"other","key":"values"}}]}}},"DocumentPostModel":{"title":"DocumentPostModel","required":["corpus_name","source_id","text","annotator","data"],"type":"object","properties":{"corpus_name":{"title":"Corpus Name","type":"string","description":"Name of the corpus where the documents belongs to"},"source_id":{"title":"Source Id","type":"integer","description":"Original ID of the Document"},"text":{"title":"Text","type":"string","description":"Text of the document"},"annotator":{"title":"Annotator","type":"string","description":"Name of the annotator"},"data":{"$ref":"#/components/schemas/AnnotationBlobModel"}},"example":{"corpus_name":"reuters","source_id":"197643","text":"Im a named entity","annotator":"gold","data":{"d_id":"some_document_id","meta":{"some":"annotator","and":"iteration","or":"creation_date"},"annotations":[{"key":"#somekey","type":"topic","surface_form":"Hello world","start":21,"end":32,"scope":"some_scope","meta":{"some":"other","key":"values"}}]}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ResponseModel":{"title":"ResponseModel","required":["status_code","message","content"],"type":"object","properties":{"status_code":{"title":"Status Code","enum":[200,201,400,401],"type":"integer","description":"Status of response"},"message":{"title":"Message","type":"string","description":"Error Message if request could not be processed"},"content":{"title":"Content","type":"object","additionalProperties":{"type":"object"}}},"example":{"status_code":200,"message":"","content":{"da_id":"some_da_id"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}